Reenable shard management on IndicesClusterStateService (#20741)
The previous commit d8b73e9 changed IndicesClusterStateService and IndicesStore to only run on master and data nodes. IndicesClusterStateService triggers acknowledgments of index deletes to the master. The current code that is waiting for the acknowledgments on the master is however expecting acknowledgements from tribe/client nodes as well. This commit reenables IndicesClusterStateService to run on tribe/client nodes to provide simple backward compatibility when acknowledging deletes but restricts the actual deletion of closed indices to data and master nodes only (while still acknowledging on tribe/client nodes). This prevents tribe/client nodes from doing unnecessary IO while looking up non-existent indices to delete.

+++ core/src/main/java/org/elasticsearch/indices/cluster/IndicesClusterStateService.java
@@ -126,17 +126,12 @@ public class IndicesClusterStateService extends AbstractLifecycleComponent<Indic
 
     @Override
     protected void doStart() {
-        // Doesn't make sense to manage shards on non-master and non-data nodes
-        if (DiscoveryNode.dataNode(settings) || DiscoveryNode.masterNode(settings)) {
-            clusterService.addFirst(this);
-        }
+        clusterService.addFirst(this);
     }
 
     @Override
     protected void doStop() {
-        if (DiscoveryNode.dataNode(settings) || DiscoveryNode.masterNode(settings)) {
-            clusterService.remove(this);
-        }
+        clusterService.remove(this);
     }
 
     @Override
@@ -230,6 +225,7 @@ public class IndicesClusterStateService extends AbstractLifecycleComponent<Indic
             }
         }
 
+        DiscoveryNode localNode = event.state().nodes().localNode();
         for (String index : event.indicesDeleted()) {
             if (logger.isDebugEnabled()) {
                 logger.debug("[{}] cleaning index, no longer part of the metadata", index);
@@ -243,7 +239,10 @@ public class IndicesClusterStateService extends AbstractLifecycleComponent<Indic
                 final IndexMetaData metaData = previousState.metaData().index(index);
                 assert metaData != null;
                 indexSettings = metaData.getSettings();
-                indicesService.deleteClosedIndex("closed index no longer part of the metadata", metaData, event.state());
+                // only delete closed indices on data / master nodes
+                if (localNode.isMasterNode() || localNode.isDataNode()) {
+                    indicesService.deleteClosedIndex("closed index no longer part of the metadata", metaData, event.state());
+                }
             }
             try {
                 nodeIndexDeletedAction.nodeIndexDeleted(event.state(), index, indexSettings, localNodeId);
